+++ tests/enums_test/subcategory_type_test.py
@@ -8,10 +8,10 @@ from funpayparsers.types.enums import SubcategoryType
 @pytest.mark.parametrize(
     'url,expected_value',
     [
-        ('https://funpay.com/en/lots/3031/', SubcategoryType.COMMON),
-        ('https://funpay.com/lots/3031/', SubcategoryType.COMMON),
-        ('https://funpay.com/en/chips/125/', SubcategoryType.CURRENCY),
-        ('https://funpay.com/chips/125/', SubcategoryType.CURRENCY),
+        ('https://funpay.com/en/lots/3031/', SubcategoryType.OFFERS),
+        ('https://funpay.com/lots/3031/', SubcategoryType.OFFERS),
+        ('https://funpay.com/en/chips/125/', SubcategoryType.CHIPS),
+        ('https://funpay.com/chips/125/', SubcategoryType.CHIPS),
         ('https://funpay.com/en/unknown_type/123/', SubcategoryType.UNKNOWN),
         ('https://funpay.com/unknown_type/123/', SubcategoryType.UNKNOWN),
     ]
@@ -23,8 +23,8 @@ def test_subcategory_type_by_url_determination(url, expected_value):
 @pytest.mark.parametrize(
     'url,expected_value',
     [
-        ('lot-3031', SubcategoryType.COMMON),
-        ('chip-128', SubcategoryType.CURRENCY),
+        ('lot-3031', SubcategoryType.OFFERS),
+        ('chip-128', SubcategoryType.CHIPS),
         ('unknown-12345', SubcategoryType.UNKNOWN),
     ]
 )