+++ tests/parsers_test/categories_parser_test.py
@@ -30,14 +30,14 @@ def single_category_obj():
             raw_source='',
             name='Аккаунты',
             id=1407,
-            type=SubcategoryType.COMMON,
+            type=SubcategoryType.OFFERS,
             offers_amount=None
         ),
         Subcategory(
             raw_source='',
             name='Ключи',
             id=1408,
-            type=SubcategoryType.COMMON,
+            type=SubcategoryType.OFFERS,
             offers_amount=None
         )
     )
@@ -78,14 +78,14 @@ def multiple_categories_obj():
             raw_source='',
             name='Кинары',
             id=6,
-            type=SubcategoryType.CURRENCY,
+            type=SubcategoryType.CHIPS,
             offers_amount=None
         ),
         Subcategory(
             raw_source='',
             name='Аккаунты',
             id=24,
-            type=SubcategoryType.COMMON,
+            type=SubcategoryType.OFFERS,
             offers_amount=None
         )
     )
@@ -95,14 +95,14 @@ def multiple_categories_obj():
             raw_source='',
             name='Кинары2',
             id=26,
-            type=SubcategoryType.CURRENCY,
+            type=SubcategoryType.CHIPS,
             offers_amount=None
         ),
         Subcategory(
             raw_source='',
             name='Аккаунты2',
             id=44,
-            type=SubcategoryType.COMMON,
+            type=SubcategoryType.OFFERS,
             offers_amount=None
         )
     )